LapinRobot-master/LapinRobot/requirements.txt
pyserial~=3.5
matplotlib~=3.5.1
pyqt5~=5.15.6
backend~=0.2.4.1
frontend~=0.0.3
pandas~=1.4.0
PyYAML~=6.0